utilisation System.Numerics

+++ hw3/Helpers/Extensions.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Numerics;
 using System.Text;
 using System.Threading.Tasks;
 
@@ -8,9 +9,9 @@ namespace hw3
 {
     public static class Extensions
     {
-        public static double ToRadians(this double deg)
+        public static float ToRadians(this float deg)
         {
-            return deg * Math.PI / 180.0d;
+            return deg * (float)Math.PI / 180.0f;
         }
     }
 }
