+++ hw3/Primitives/GeometricPrimitive.cs
@@ -21,7 +21,7 @@ namespace hw3
             return Material.GetShading(geo);
         }
 
-        public bool Intersect(Ray ray, bool computeGeo, out LocalGeo geo, out double t)
+        public bool Intersect(Ray ray, bool computeGeo, out LocalGeo geo, out float t)
         {
             return Shape.Intersect(ray, computeGeo, out geo, out t);
         }
